Clean-up xUnit file structure and obsolete code.
