+++ test/EFCore.MySql.FunctionalTests/TestUtilities/Xunit/MySqlConditionalFactDiscoverer.cs
@@ -1,0 +1,24 @@
+﻿using Microsoft.EntityFrameworkCore.TestUtilities.Xunit;
+using Xunit.Abstractions;
+using Xunit.Sdk;
+
+namespace Pomelo.EntityFrameworkCore.MySql.FunctionalTests.TestUtilities.Xunit
+{
+    public class MySqlConditionalFactDiscoverer : ConditionalFactDiscoverer
+    {
+        public MySqlConditionalFactDiscoverer(IMessageSink messageSink)
+            : base(messageSink)
+        {
+        }
+
+        protected override IXunitTestCase CreateTestCase(
+            ITestFrameworkDiscoveryOptions discoveryOptions,
+            ITestMethod testMethod,
+            IAttributeInfo factAttribute)
+            => new MySqlConditionalFactTestCase(
+                DiagnosticMessageSink,
+                discoveryOptions.MethodDisplayOrDefault(),
+                discoveryOptions.MethodDisplayOptionsOrDefault(),
+                testMethod);
+    }
+}
