+++ test/EFCore.MySql.FunctionalTests/TestUtilities/Xunit/MySqlConditionalTheoryDiscoverer.cs
@@ -1,0 +1,41 @@
+﻿using System.Collections.Generic;
+using Microsoft.EntityFrameworkCore.TestUtilities.Xunit;
+using Xunit.Abstractions;
+using Xunit.Sdk;
+
+namespace Pomelo.EntityFrameworkCore.MySql.FunctionalTests.TestUtilities.Xunit
+{
+    public class MySqlConditionalTheoryDiscoverer : ConditionalTheoryDiscoverer
+    {
+        public MySqlConditionalTheoryDiscoverer(IMessageSink messageSink)
+            : base(messageSink)
+        {
+        }
+
+        protected override IEnumerable<IXunitTestCase> CreateTestCasesForTheory(
+            ITestFrameworkDiscoveryOptions discoveryOptions,
+            ITestMethod testMethod,
+            IAttributeInfo theoryAttribute)
+        {
+            yield return new MySqlConditionalTheoryTestCase(
+                DiagnosticMessageSink,
+                discoveryOptions.MethodDisplayOrDefault(),
+                discoveryOptions.MethodDisplayOptionsOrDefault(),
+                testMethod);
+        }
+
+        protected override IEnumerable<IXunitTestCase> CreateTestCasesForDataRow(
+            ITestFrameworkDiscoveryOptions discoveryOptions,
+            ITestMethod testMethod,
+            IAttributeInfo theoryAttribute,
+            object[] dataRow)
+        {
+            yield return new MySqlConditionalFactTestCase(
+                DiagnosticMessageSink,
+                discoveryOptions.MethodDisplayOrDefault(),
+                discoveryOptions.MethodDisplayOptionsOrDefault(),
+                testMethod,
+                dataRow);
+        }
+    }
+}
