+++ test/EFCore.MySql.FunctionalTests/TestUtilities/Xunit/MySqlXunitTestFramework.cs
@@ -1,0 +1,15 @@
+﻿using Xunit.Abstractions;
+using Xunit.Sdk;
+
+namespace Pomelo.EntityFrameworkCore.MySql.FunctionalTests.TestUtilities.Xunit
+{
+    public class MySqlXunitTestFramework : XunitTestFramework
+    {
+        public MySqlXunitTestFramework(IMessageSink messageSink) : base(messageSink)
+        {
+        }
+
+        protected override ITestFrameworkDiscoverer CreateDiscoverer(IAssemblyInfo assemblyInfo)
+            => new MySqlXunitTestFrameworkDiscoverer(assemblyInfo, SourceInformationProvider, DiagnosticMessageSink);
+    }
+}
